requirements.txt
apache-log-parser==1.7.0
Faker==0.7.3
ipaddress==1.0.17
numpy==1.11.2
python-dateutil==2.5.3
pytz==2016.7
six==1.10.0
ua-parser==0.7.1
user-agents==1.0.1
